packages/daemon/pyproject.toml
[tool.poetry]
name = "orchestra-daemon"
version = "0.1.0"
description = "Project Orchestra FastAPI daemon"
authors = ["Dean <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
PyJWT = { version = "^2.9.0", extras = ["crypto"] }
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.7"
coverage = "^7.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
